requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
black==22.10.0
cachetools==4.2.2
certifi==2022.9.24
click==8.1.3
colorama==0.4.6
Django==4.1.3
django-debug-toolbar==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mypy-extensions==0.4.3
pathspec==0.10.2
platformdirs==2.5.4
PyJWT==2.6.0
python-dotenv==0.21.0
python-telegram-bot==13.14
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tornado==6.1
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
stripe==5.0.0